Adds support for dynamic handler notification
The keypair generated by this role will surely be associated with a
service of some kind. If swapping the cert should notify a handler,
simply declare the handler(s) via the list var. Provided the parent
role ships with the named handler, this role will run it without issue.

By default, of course, no handlers are notified, since the var value
defaults to an empty list.